remove unnecessary test
